+++ src/main.rs
@@ -5,8 +5,7 @@
     - Welcome to Tron's source code
 
 
-    latest version: 2.7.0
-    latest release: Feb 14
+    latest version: 2.8.0 (Feb 22, 2024)
 
 */
 mod environment;
@@ -21,10 +20,7 @@ use crate::interpreter::*;
 use crate::parser::*;
 use crate::resolver::*;
 use crate::scanner::*;
-use std::env;
-use std::fs;
-use std::process::exit;
-use std::process::Command;
+use std::{env, fs, process::{exit, Command}};
 
 pub fn panic(message: &str) {
     eprintln!("\x1B[31m{}\x1B[0m \n", message);
@@ -32,21 +28,16 @@ pub fn panic(message: &str) {
 }
 
 pub fn run_file(path: &str) -> Result<(), String> {
-    let absolute_path = if path.starts_with("/") {
-        path.to_string()
-    } else {
-        let current_dir = std::env::current_dir().unwrap();
-        current_dir.join(path).to_str().unwrap().to_string()
-    };
-    match fs::read_to_string(&absolute_path) {
+    let current_dir = std::env::current_dir().unwrap();
+    match fs::read_to_string(&current_dir.join(path).to_str().unwrap().to_string()) {
         Err(msg) => Err(msg.to_string()),
-        Ok(contents) => run_string(&contents),
+        Ok(contents) => {
+            let mut interpreter = Interpreter::new();
+            run(&mut interpreter, &contents)
+        }
     }
 }
-pub fn run_string(contents: &str) -> Result<(), String> {
-    let mut interpreter = Interpreter::new();
-    run(&mut interpreter, contents)
-}
+
 fn run(interpreter: &mut Interpreter, contents: &str) -> Result<(), String> {
     let scanner = Scanner::new(contents);
     let tokens = scanner.scan_tokens()?;
@@ -56,16 +47,21 @@ fn run(interpreter: &mut Interpreter, contents: &str) -> Result<(), String> {
     let locals = resolver.resolve(&stmts.iter().collect())?;
     interpreter.resolve(locals);
     interpreter.interpret(stmts.iter().collect())?;
-    return Ok(());
+    Ok(())
 }
+
 fn main() {
     let args: Vec<String> = env::args().collect();
     let path = std::env::current_dir().unwrap();
-    if args.len() == 2 {
-        let command = &args[1];
-        if command == "version" {
-            println!("v2.7.0")
-        } else if command == "update" {
+    if args.len() == 1 {
+        exit(64);
+    }
+    let command = args[1].as_str();
+    match command {
+        "version" => {
+            println!("v2.8.0");
+        }
+        "update" => {
             println!("Updating....");
             let _output = Command::new("bash")
                 .arg("-c")
@@ -73,9 +69,20 @@ fn main() {
                 .output()
                 .expect("Failed to execute command");
             println!("Update completed");
-        } else {
-            let filename = command;
-            let path_buf = path.join(filename);
+        }
+        "help" => {
+            println!(
+                "
+        \x1B[32mtron\x1B[0m \x1B[33m<filename>\x1B[0m - interpret tron files
+        \x1B[32mtron\x1B[0m \x1B[34mversion\x1B[0m - current version of Tron
+        \x1B[32mtron\x1B[0m \x1B[34mupdate\x1B[0m - install latest version of Tron
+
+             \x1B[38;5;208mTron Programming Language (2.8.0)\x1B[0m
+        "
+            );
+        }
+        _ => {
+            let path_buf = path.join(command);
             let input = path_buf.to_str();
             match input {
                 Some(input) => match run_file(input) {
@@ -90,16 +97,5 @@ fn main() {
                 }
             }
         }
-    } else {
-        println!(
-            "
-        \x1B[32mtron\x1B[0m \x1B[33m<filename>\x1B[0m - interpret tron files
-        \x1B[32mtron\x1B[0m \x1B[34mversion\x1B[0m - current version of Tron
-        \x1B[32mtron\x1B[0m \x1B[34mupdate\x1B[0m - install latest version of Tron
-
-             \x1B[38;5;208mTron Programming Language (2.7.0)\x1B[0m
-        "
-        );
-        exit(64);
     }
 }